Merged PR 21340: WinHttpConnect accepts Unicode not ASCII
WinHttpConnect accepts Unicode not ASCII

Existing text said "the IP address of the site in ASCII", but the string is a Unicode string. This is a leftover from the internal implementation, which calls all strings "ASCII" even when they are Unicode.

The incorrect terminology has confused developers, who were under the mistaken impression that they had to pass an 8-bit string.

https://stackoverflow.com/q/78821597