Merged PR 21358: Fix HTML escapes in syntax blocks (#5)
These are the last remaining topics in the repro with HTML escape issues in syntax blocks.

Sorry, I didn't break this final batch up by associated author.  The topics are scattered across a variety of you, but the change is a simple mechanical search and replace and should be reviewable by anybody.