Merged PR 21362: d2d1 : Cleanup weird replacement characters
These topics have a weird placeholder (�) character that shouldn't be there.

For example, here:
https://learn.microsoft.com/en-us/windows/win32/api/d2d1_1/nf-d2d1_1-id2d1properties-getpropertyname(u_pwstr_uint32)

There is one following the HRESULT table in the Return value section, they appear in the middle of "Windows 8" etc. in Requirement section, and there are a couple at the end of the See also list.

In the Requirements section, I replaced these with non-breaking spaces.  Unclear if non-breaking spaces are really necessary here, but they are used in nearly all the other topics in the repro so this is just being consistent.

Just deleted the others.