Merged PR 21369: Correct IDCompositionDevice::CreateRectangleClip remarks
I looked at these initially because of the broken FLT_MAX links.  However, the original comments were not even correct, the default values of these properties are not relative to FLT_MAX but instead are bounded by -2^21 to 2^21.

No symbolic constants was ever defined for these limits, so we just use the literal numeric values.

IDCompositionRectangleClip::SetBottom et. al. reference the same constants as the upper/lower bounds of the clip rectangle.  I also looked at the underlying product code to verify that this was correct.

@<Steven White> FYI.